refactor: reorganizado pacotes do projeto agrupando entidades por dominio
