+++ src/main/java/com/api/barber_shop/domain/repositories/BarberRepository.java
@@ -1,21 +1,22 @@
-package com.api.barber_shop.repositorys;
+package com.api.barber_shop.domain.repositories;
 
-import com.api.barber_shop.entity.BarberEntity;
+import com.api.barber_shop.domain.entities.BarberEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 @Repository
 public interface BarberRepository extends JpaRepository<BarberEntity, UUID> {
 
-    //Containg busca parcialmente(contém dentro) e ignora maiuscula e minuscula na comparação.
+
     List<BarberEntity> findByNameBarberContainingIgnoreCase(String nameBarber);
 
     List<BarberEntity> findBySpecialtyBarberContainingIgnoreCase(String specialtyBarber);
 
-    //Buscas pelos dois(name/especialidade)
     List<BarberEntity> findByNameBarberContainingIgnoreCaseAndSpecialtyBarberContainingIgnoreCase(String name, String epecialty);
+
+
+    List<BarberEntity> id(UUID id);
 }
