+++ src/main/java/com/api/barber_shop/dtos/BarberResponseDto.java
@@ -1,6 +1,6 @@
 package com.api.barber_shop.dtos;
 
-import com.api.barber_shop.entity.enums.ActiveBarber;
+import com.api.barber_shop.domain.entities.enums.ActiveBarber;
 
 import java.util.UUID;
